refactor: 모달 닫기 버튼 컴포넌트 Modal 컴포넌트 폴더로 이동
