+++ src/components/_commons/Modal/Modal.js
@@ -3,7 +3,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import { ModalSection, ModalInnerDiv } from './Modal.style';
-import { ModalCloseButton } from '../../CloseButton/CloseButton';
+import { ModalCloseButton } from './ModalCloseButton/ModalCloseButton';
 
 export const Modal = props => {
   const { children, onClick, ...rest } = props;
